Comments + Debugging

+++ Lab3DynamicWebHibernateAnnotations/WebContent/Student/deleteStudent.jsp
@@ -7,7 +7,8 @@
 <%@ page import="java.io.PrintWriter"%>
 <%@ page import="java.util.ArrayList"%>
 <%@ page import="md.victordov.lab.dao.*"%>
-
+<!DOCTYPE html>
+<!-- Delete Student page receives id through POST METHOD and check if not null, in case a null id a blank page will be displayed -->
 <html>
 <head>
 <link href="<%=request.getContextPath()%>/style.css" rel="stylesheet"
@@ -15,13 +16,16 @@
 <title>Student Delete</title>
 </head>
 <body>
-
+	<%@ include file="/headerJSP.jsp"%>
 	<%
-		GenericDAO<Student> genDao = new StudentDAO();
-		Student stud = new Student();
-		String idStudentString = request.getParameter("id");
-		Integer idStudent = Integer.parseInt(idStudentString);
-		genDao.delete(idStudent);
+		if ("POST".equalsIgnoreCase(request.getMethod())
+				&& (request.getParameter("id") != null)) {
+			GenericDAO<Student> genDao = new StudentDAO();
+			Student stud = new Student();
+			String idStudentString = request.getParameter("id");
+			Integer idStudent = Integer.parseInt(idStudentString);
+			genDao.delete(idStudent);
+		}
 	%>
 	Click Here:
 	<a href="<%=request.getContextPath()%>/Student/StudentJSP.jsp"><strong>Student
