+++ Lab3DynamicWebHibernateAnnotations/WebContent/Universitate/deleteUniv.jsp
@@ -7,7 +7,9 @@
 <%@ page import="java.io.PrintWriter"%>
 <%@ page import="java.util.List"%>
 <%@ page import="md.victordov.lab.dao.*"%>
-
+<!DOCTYPE html>
+<!-- Page that will delete the object of class Univesitate, through the Post Method id is sent and verified if not null.
+If id is sent through the browser bar or directly a black page will be displayed and void action is taken-->
 <html>
 <head>
 <link href="<%=request.getContextPath()%>/style.css" rel="stylesheet"
@@ -20,14 +22,17 @@
 	<%@ include file="/headerJSP.jsp"%>
 
 	<%
-		GenericDAO<Universitate> genDao = new UnivDAO();
-		Universitate univ = new Universitate();
-		String idUniverString = request.getParameter("idUniversitate");
-		Integer idUniversitate = Integer.parseInt(idUniverString);
-		genDao.delete(idUniversitate);
+		if ("POST".equalsIgnoreCase(request.getMethod())
+				&& (request.getParameter("idUniversitate") != null)) {
+			GenericDAO<Universitate> genDao = new UnivDAO();
+			Universitate univ = new Universitate();
+			String idUniverString = request.getParameter("idUniversitate");
+			Integer idUniversitate = Integer.parseInt(idUniverString);
+			genDao.delete(idUniversitate);
+		}
 	%>
-	<a href="<%=request.getContextPath()%>/Universitate/UnivJSP.jsp">Apasaaici:
-		<strong>Universitate</strong>
+	Click here:
+	<a href="<%=request.getContextPath()%>/Universitate/UnivJSP.jsp"> <strong>Universitate</strong>
 	</a>
 
 	<!-- Footer -->
